Fix cloned handlers to use same mutex (#97)
* test: Test cloned handlers synchronize writer writers

All handlers cloned from the handler attached to a logger should synchronize writer writes. This allows the logger to be used from multiple go routines.

Here we add a test writing to a `bytes.Buffer` which itself is in not synchronized to expose the problem.

* fix: Cloned handlers use the same mutex

Drawing on the information in https://github.com/golang/example/blob/8b405629c4a5215871be932097e099c05ec5cb2e/slog-handler-guide/README.md#concurrency-safety, the handler is adapted to store a pointer to a mutex so that it and all of its clones synchronize with each other.

* workflows: test with `-race`

* added comment

* small cleanup

---------

Co-authored-by: lmittmann <lmittmann@users.noreply.github.com>